requirements.txt
tornado==6.3.1
numpy==1.24.3
python-docx==0.8.11
Pillow==9.5.0